renamed 'source' folder to 'project' to allow nosetests to run
