+++ project/static/main.js
