+++ project/templates/user/activate.html
