+++ project/user/forms.py
@@ -5,7 +5,7 @@ from flask_wtf import Form
 from wtforms import StringField, PasswordField
 from wtforms.validators import DataRequired, Email, Length, EqualTo
 
-from source.models import User
+from project.models import User
 
 
 class LoginForm(Form):
